Fix completion candidates for KUBEL-SET-CONTEXT (#103)
* Fix completion candidates for KUBEL-SET-CONTEXT

We can't use SHELL-COMMAND-TO-STRING as calls SHELL-COMMAND with ERROR-BUFFER as
nil. This mixes output from stdout and stderr. This can be an issue as kubectl
will write warnings about the SSO session having expired to standard error, even
when the command can run successfully to w/o an valid session. It still returns a
successful error code.

The specific scenario I run into is when I call KUBEL-SET-CONTEXT w/o a valid
session. The call to retrieve the available contexts expects the shell command
to return a space separated list of contexts. Because stdout and stderr are
combined that list also includes the error message. This results in the call to
completing read to have a lot of nonsense entries, one for each word the error
message has.

* Fix spec

As KUBEL--EXEC-TO-STRING no longer calls SHELL-COMMAND-TO-STRING so we need to
patch KUBEL--EXEC-TO-STRING directly to return the version string.