Remove kubectl version check (#117)
I noticed a check that verifies if the server version is greater than 1.13. As per the Kubernetes
release documentation (https://kubernetes.io/releases/patch-releases/), version 1.13 has reached its
end-of-life (EOL) status since 2019. Considering this information, I strongly believe that this
check is unnecessary and can be safely removed.

It's only used here:
- https://github.com/abrochard/kubel/blob/a84f09cca8cdc71d55aa78304db4156df46d04e8/kubel.el#L896

Removing this check will not impact the functionality of the codebase, as it is only required for
outdated Kubernetes versions that are more than four years old and no longer supported. It is highly
unlikely that anyone would be using such an outdated version that cannot be found online for
download.

By removing this unnecessary code, we can eliminate technical debt and improve the overall
readability and maintainability of the codebase.

Thank you for considering this proposal.