fix: Don't assume that kubectl will use the namespace default implicitely
When kubel-namespace is default, the code explicitely avoided providing -n
default in the command line. This resulted in kubectl using the value in the
kube configuration and falling back to default.

In the former case, having kubel-namespace set to default did use the configured
value, leading to a counter intuitive behavior.